+++ src/defRepo/check.test.ts
@@ -5,13 +5,13 @@ import { Response } from 'node-fetch';
 import path from 'path';
 
 import { getAbsolutePathToRepositoryRoot, getGitSha } from '../vcs/git';
-import { executeApiCall, ApiCallArguments } from '../api';
+import { executeApiCall, ApiCallArguments } from '../api/http';
 import { getGitHubRepositoryName } from '../vcs/github';
 import { executeCheck } from './check';
 
 jest.mock('../vcs/github');
 jest.mock('../vcs/git');
-jest.mock('../api');
+jest.mock('../api/http');
 
 const executeApiCallMock = executeApiCall as jest.Mock;
 const getGitShaMock = getGitSha as jest.Mock;