+++ src/serviceDefinition/exec/taskIdentification.ts
@@ -1,13 +1,11 @@
 // SPDX-License-Identifier: MIT
 // SPDX-FileCopyrightText: 2022 Steadybit GmbH
 
-import { ServiceDefinition, Task } from '../types';
+import { ServiceState, Task } from '../types';
 import { Options } from './types';
-import { getState } from '../api';
 
-export async function getTasks(options: Options, serviceDefinition: ServiceDefinition): Promise<Task[]> {
-  const state = await getState(serviceDefinition);
-  let tasks = state.tasks;
+export function getTasks(options: Options, serviceState: ServiceState): Task[] {
+  let tasks = serviceState.tasks;
 
   const taskFilter = options.task;
   if (taskFilter != null) {