+++ src/serviceDefinition/show.test.ts
@@ -2,11 +2,11 @@
 // SPDX-FileCopyrightText: 2022 Steadybit GmbH
 
 import { loadServiceDefinition } from './files';
-import { executeApiCall } from '../api';
+import { executeApiCall } from '../api/http';
 import { executeShow } from './show';
 
 jest.mock('./files');
-jest.mock('../api');
+jest.mock('../api/http');
 
 const loadServiceDefinitionMock = loadServiceDefinition as jest.Mock;
 const executeApiCallMock = executeApiCall as jest.Mock;