+++ src/serviceDefinition/types.d.ts
@@ -52,13 +52,23 @@ export type Parameters = Record<string, any>;
 
 export type TaskState = 'PENDING' | 'SUCCESS' | 'FAILURE';
 
+export interface TaskHistoryEntry {
+  date: string;
+  type: string;
+  meta: Record<string, any>;
+}
+
 export interface Task {
+  id: string;
   definition: ReferenceCoordinate;
   state: TaskState;
+  type: string;
+  history: TaskHistoryEntry[];
   parameters: Parameters;
   matrixContext: never;
 }
 
 export interface ServiceState {
+  id: string;
   tasks: Task[];
 }